Remove blocked domain service and its receiver
Reported blocked domains can be done directly without registering it as a service.
BlockedDomainAlarmReceiver.java doesn't do anything anyway as it was disabled to avoid continuous FC.
In fact, processing reported blocked domains is the same anyway whether it is using a service or not.
So, better avoid it to not causing more overhead.