Fix whitelist URL for a specific app bug
Apparently, whitelisting an URL for a specific app doesn't work as expected.
It behaves the same as whitelisting the whole app which is not desired.

I thought, the deny list doesn't need to be filled as there is a rule anyway to block domains for all apps.
This is unfortunately, not true. The blocked domains list needs also to filled into the rule.
Because of this changes, I need to refactor ContentBlocker56 to suit best with this behavior.